third_party/WebKit/LayoutTests/fast/js/cross-frame-prototype.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<iframe></iframe>
<script>

    var arrayConstructor = frames[0].Array;
    arrayConstructor.prototype.foo = "bar";
    var x = new arrayConstructor();

    shouldBe("x.foo", '"bar"');
</script>
</body>
</html>
